+++ web/app/mocks/db/index.ts
@@ -1,4 +1,5 @@
-import { IndicatorInfoResponse, IndicatorValueResponse } from '../../store/querys/numerical-guidance/indicator.query';
+import { IndicatorValueResponse } from '../../store/querys/numerical-guidance/indicator.query';
+import { IndicatorInfoResponse } from '@/app/store/querys/numerical-guidance/indicator-list.query';
 import { IndicatorBoardMetadataResponse } from '../../store/querys/numerical-guidance/indicator-board-metadata.query';
 import { indicatorsValueMockData } from '../mock-data/indicators-value.mock';
 import { CustomForecastIndicatorListResponse } from '../../store/querys/numerical-guidance/custom-forecast-indicator.query';
