+++ web/app/store/querys/numerical-guidance/indicator.query.ts
@@ -1,15 +1,9 @@
 import useSWR from 'swr';
 import { API_PATH } from '../api-path';
-import { defaultFetcher, fetchIndicatorsValue } from '../fetcher';
+import { fetchIndicatorsValue } from '../fetcher';
 import { Interval } from '../../stores/numerical-guidance/indicator-board.store';
 
 export type IndicatorType = 'customForecastIndicator' | 'k-stock';
-
-export type IndicatorInfoResponse = {
-  id: string;
-  ticker: string;
-  name: string;
-};
 
 export type IndicatorsValueResponse = {
   indicatorsValue: IndicatorValueResponse[];
@@ -27,8 +21,6 @@ export type IndicatorValueItemResponse = {
   date: string;
   value: number | string;
 };
-
-export const useFetchIndicatorList = () => useSWR<IndicatorInfoResponse[]>(API_PATH.indicatorList, defaultFetcher);
 
 export const useFetchLiveIndicatorsValue = (indicatorIds: string[] | undefined, interval: Interval = 'day') => {
   // fix: type에 따라 url 변경 필요
