+++ web/app/ui/components/numerical-guidance/custom-forecast-indicator/source-indicator-slider.tsx
@@ -1,4 +1,4 @@
-import { IndicatorInfoResponse } from '@/app/store/querys/numerical-guidance/indicator.query';
+import { IndicatorInfoResponse } from '@/app/store/querys/numerical-guidance/indicator-list.query';
 import { Card } from '@tremor/react';
 import Slider from '../../view/atom/slider/slider';
 import { useSelectedCustomForecastIndicatorViewModel } from '@/app/business/hooks/custom-forecast-indicator/use-selected-custom-forecast-indicator-view-model';
