+++ web/app/ui/components/numerical-guidance/indicator/indicator-list/indicator-list.tsx
@@ -4,7 +4,7 @@ import { useIndicatorList } from '@/app/business/hooks/indicator/use-indicator-l
 import React from 'react';
 import { useSelectedIndicatorBoardMetadata } from '@/app/business/hooks/indicator-board-metedata/use-selected-indicator-board-metadata-view-model.hook';
 import IndicatorListItem from './indicator-list-item';
-import { IndicatorInfoResponse } from '@/app/store/querys/numerical-guidance/indicator.query';
+import { IndicatorInfoResponse } from '@/app/store/querys/numerical-guidance/indicator-list.query';
 import { cn } from '@/app/utils/style';
 
 const IndicatorList = React.memo(function IndicatorList() {
